using usize for limit

+++ crates/mimic/src/data/query/load/dynamic.rs
@@ -84,8 +84,8 @@ pub struct LoadQueryDyn {
     pub selector: Selector,
     pub format: LoadFormat,
     pub lookup: Option<Where>,
-    pub limit: Option<u32>,
-    pub offset: u32,
+    pub limit: Option<usize>,
+    pub offset: usize,
     pub include_children: bool,
 }
 
@@ -114,21 +114,21 @@ impl LoadQueryDyn {
 
     // offset
     #[must_use]
-    pub const fn offset(mut self, offset: u32) -> Self {
+    pub const fn offset(mut self, offset: usize) -> Self {
         self.offset = offset;
         self
     }
 
     // limit
     #[must_use]
-    pub const fn limit(mut self, limit: u32) -> Self {
+    pub const fn limit(mut self, limit: usize) -> Self {
         self.limit = Some(limit);
         self
     }
 
     // limit_option
     #[must_use]
-    pub const fn limit_option(mut self, limit: Option<u32>) -> Self {
+    pub const fn limit_option(mut self, limit: Option<usize>) -> Self {
         self.limit = limit;
         self
     }
